- Reorganize junit tests
